+++ tests/assertions/assert_text.test.js
@@ -29,10 +29,19 @@ describe("Assert Text", function() {
         await browser.assert.text(node, "Main Title");
     });
 
+    it("Text From Xpath", async () => {
+        await browser.assert.text("//h1", "Main Title");
+    });
+
     it("Multiple Texts", async () => {
         await browser.assert.text("p", "My first paragraph");
         await browser.assert.text("p", "My second paragraph");
     });
+
+    it("Text Regex", async () => {
+        await browser.assert.text("h1", /Main\sTitle/);
+    });
+
 
     it("Text Throws", async () => {
         await utils.assertThrowsAssertionAsync(async () => {
@@ -46,6 +55,12 @@ describe("Assert Text", function() {
         }, `Expected element ".container p" to have text "My second paragraph", "My first paragraph" found`);
     });
 
+    it("Text Regex Throws", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.text("h1", /My\sfirst\sparagraph/);
+        }, `Expected element "h1" to have text "/My\\sfirst\\sparagraph/", "Main Title" found`);
+    });
+
     it("Multiple Text Throws", async () => {
         await utils.assertThrowsAssertionAsync(async () => {
             await browser.assert.text("p", "My paragraph");
@@ -56,6 +71,33 @@ describe("Assert Text", function() {
         await utils.assertThrowsAssertionAsync(async () => {
             await browser.assert.text("h1", "My first paragraph", "text failed");
         }, `text failed`);
+    });
+
+    it("Not Text", async () => {
+        await browser.assert.not.text("h1", "Not Main Title");
+    });
+
+    it("Not Text Not Exists", async () => {
+        await browser.assert.not.text("h2", "Not Main Title");
+    });
+
+    it("Not Multiple Texts", async () => {
+        await browser.assert.not.text("p", "not a paragraph");
+    });
+
+    it("Not Multiple Texts Throws", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.not.text("p", "My first paragraph");
+        }, `Expected element "p" not to have text "My first paragraph"`);
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.not.text("p", "My second paragraph");
+        }, `Expected element "p" not to have text "My second paragraph"`);
+    });
+
+    it("Not Text Throws With Custom Message", async () => {
+        await utils.assertThrowsAssertionAsync(async () => {
+            await browser.assert.not.text("p", "My first paragraph", "not text failed");
+        }, `not text failed`);
     });
 
 });